# Reduce bcrypt cost and Clone