--- a/Cargo.toml
+++ b/Cargo.toml
@@ -10,6 +10,7 @@ description = "Simple file server for comic books"
 [dependencies]
 askama = "0.12.1"
 axum = "0.7.4"
+axum-extra = { version = "0.9.2", features = ["typed-header"] }
 base64 = "0.22.0"
 bcrypt = "0.15.0"
 blake3 = "1.5.1"