before

```
Running 10s test @ http://localhost:8080
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    21.63ms    3.20ms  57.07ms   94.44%
    Req/Sec   232.38     22.78   262.00     91.50%
  4630 requests in 10.01s, 401.63MB read
Requests/sec:    462.62
Transfer/sec:     40.13MB
```

after

```
Running 10s test @ http://localhost:8080
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.05ms  450.02us   8.84ms   85.43%
    Req/Sec     4.89k   361.77     8.91k    93.53%
  97884 requests in 10.10s, 8.29GB read
Requests/sec:   9691.49
Transfer/sec:    840.68MB
```